pyproject.toml
[project]
name = "code-craft-03"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2",
    "django-extensions>=4.1",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "drf-nested-routers>=0.94.1",
    "drf-yasg>=1.21.10",
    "pillow>=11.2.1",
]
